change-request-service/requirements.txt
Flask_JWT_Extended==3.24.1
psycopg2
Flask_Cors==3.0.9
Flask==1.1.2
sshtunnel==0.1.5
python-dotenv==0.15.0
flask-restplus==0.13.0
Werkzeug==0.16.1
pyjwt
